backend/requirements.txt
# backend/requirements.txt
django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
google-genai
python-dotenv==1.0.1
